Remove the big Nope from the title of the home page
I originally thought it was just a humorous expression and didn't think
much of it. However, recently I saw someone use the title on the website
to suggest that Modules are still very immature, which I feel is
inconsistent with our original intention. We hope that this website can
provide more people with a clearer understanding of modules and
encourage them to use them. But if the title is now having the opposite
effect, I think it would be better to remove it.