CSS: display "Unknown", not "Unkno...", in table (#67)
* Display the entire text on mouseover

The "truncate-text" class is redundant with "td", so remove it entirely.
Mouseover (hover) changes overflow from hidden to visible, so the user
can see the whole column entry.

* CSS: display "Unknown", not "Unkno...", in table

Steal these 30 pixels from the "import directive" column, since it
doesn't need them — and since the previous commit, it's not a disaster
even if that column does overflow, because you can see the whole entry
on hover anyway. (But, for the record, it doesn't overflow on my screen,
either.)

The word "deprecated" still displays as "deprecat..." even after
this patch.

* CSS: display "deprecated", not "deprecat...", in table

Stealing another 10 pixels from the import column actually does
make that column overflow, so, let's make overflow scroll instead
of hiding/mutating it entirely. This also means hovering isn't
so important (good because it's difficult on mobile).